Move test_utils into tests module
