+++ crates/ide_completion/src/completions/attribute/derive.rs
@@ -82,7 +82,7 @@ const DEFAULT_DERIVE_COMPLETIONS: &[DeriveDependencies] = &[
 mod tests {
     use expect_test::{expect, Expect};
 
-    use crate::{test_utils::completion_list, CompletionKind};
+    use crate::{tests::filtered_completion_list, CompletionKind};
 
     fn check(ra_fixture: &str, expect: Expect) {
         let builtin_derives = r#"
@@ -106,7 +106,7 @@ pub macro PartialOrd {}
 pub macro Ord {}
 
 "#;
-        let actual = completion_list(
+        let actual = filtered_completion_list(
             &format!("{} {}", builtin_derives, ra_fixture),
             CompletionKind::Attribute,
         );
